Add URL datasource
Adds a URL based datasource retrieved with go-resty. Reworks the
existing ConfigMap API, and Controller code so that resources can be
correctly deleted and updated. This does mean that a ConfigMap or URL
will be retrieved twice if it has changed (once by Observe, and once by
Create or Update).

URL datasources must return JSON and will timeout in 1 second, so if the
data takes a long time to retrieve then it must be cached or
pre-generated on the remote end.

Signed-off-by: Ben Agricola <bagricola@squiz.co.uk>